+++ Database/MongoDB/Internal/Protocol.hs
@@ -8,10 +8,18 @@
 {-# LANGUAGE CPP, FlexibleContexts, TupleSections, TypeSynonymInstances #-}
 {-# LANGUAGE MultiParamTypeClasses, FlexibleInstances, UndecidableInstances #-}
 
+{-# LANGUAGE NamedFieldPuns, ScopedTypeVariables #-}
+
+#if (__GLASGOW_HASKELL__ >= 706)
+{-# LANGUAGE RecursiveDo #-}
+#else
+{-# LANGUAGE DoRec #-}
+#endif
+
 module Database.MongoDB.Internal.Protocol (
     FullCollection,
     -- * Pipe
-    Pipe, newPipe, newPipeWith, send, call,
+    Pipe, newPipe, send, call,
     -- ** Notice
     Notice(..), InsertOption(..), UpdateOption(..), DeleteOption(..), CursorId,
     -- ** Request
@@ -19,7 +27,8 @@ module Database.MongoDB.Internal.Protocol (
     -- ** Reply
     Reply(..), ResponseFlag(..),
     -- * Authentication
-    Username, Password, Nonce, pwHash, pwKey
+    Username, Password, Nonce, pwHash, pwKey,
+    isClosed, close
 ) where
 
 #if !MIN_VERSION_base(4,8,0)
@@ -35,7 +44,15 @@ import Data.IORef (IORef, newIORef, atomicModifyIORef)
 import System.IO (Handle)
 import System.IO.Unsafe (unsafePerformIO)
 import Data.Maybe (maybeToList)
+import GHC.Conc (ThreadStatus(..), threadStatus)
+import Control.Monad (forever)
+import Control.Concurrent.Chan (Chan, newChan, readChan, writeChan)
+import Control.Concurrent (ThreadId, forkIO, killThread)
 
+import Control.Exception.Lifted (onException, throwIO, try)
+import qualified Control.Exception.Lifted as CEL
+
+import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as L
 
 import Control.Monad.Trans (MonadIO, liftIO)
@@ -49,37 +66,104 @@ import qualified Data.Text as T
 import qualified Data.Text.Encoding as TE
 
 import Database.MongoDB.Internal.Util (whenJust, bitOr, byteStringHex)
-import System.IO.Pipeline (Pipeline, newPipeline, IOStream(..))
+import System.IO (hClose, hFlush)
 
-import qualified System.IO.Pipeline as P
+#if MIN_VERSION_base(4,6,0)
+import Control.Concurrent.MVar.Lifted (MVar, newEmptyMVar, newMVar, withMVar,
+                                       putMVar, readMVar, mkWeakMVar)
+#else
+import Control.Concurrent.MVar.Lifted (MVar, newEmptyMVar, newMVar, withMVar,
+                                         putMVar, readMVar, addMVarFinalizer)
+#endif
 
-import Database.MongoDB.Internal.Connection (Connection)
-import qualified Database.MongoDB.Internal.Connection as Connection
+#if !MIN_VERSION_base(4,6,0)
+mkWeakMVar :: MVar a -> IO () -> IO ()
+mkWeakMVar = addMVarFinalizer
+#endif
+
+-- * Pipeline
+
+-- | Thread-safe and pipelined connection
+data Pipeline = Pipeline {
+    vStream :: MVar Handle,  -- ^ Mutex on handle, so only one thread at a time can write to it
+    responseQueue :: Chan (MVar (Either IOError Response)),  -- ^ Queue of threads waiting for responses. Every time a response arrive we pop the next thread and give it the response.
+    listenThread :: ThreadId
+    }
+
+-- | Create new Pipeline over given handle. You should 'close' pipeline when finished, which will also close handle. If pipeline is not closed but eventually garbage collected, it will be closed along with handle.
+newPipeline :: Handle -> IO Pipeline
+newPipeline stream = do
+    vStream <- newMVar stream
+    responseQueue <- newChan
+    rec
+        let pipe = Pipeline{..}
+        listenThread <- forkIO (listen pipe)
+    _ <- mkWeakMVar vStream $ do
+        killThread listenThread
+        hClose stream
+    return pipe
+
+close :: Pipeline -> IO ()
+-- ^ Close pipe and underlying connection
+close Pipeline{..} = do
+    killThread listenThread
+    hClose =<< readMVar vStream
+
+isClosed :: Pipeline -> IO Bool
+isClosed Pipeline{listenThread} = do
+    status <- threadStatus listenThread
+    return $ case status of
+        ThreadRunning -> False
+        ThreadFinished -> True
+        ThreadBlocked _ -> False
+        ThreadDied -> True
+--isPipeClosed Pipeline{..} = isClosed =<< readMVar vHandle  -- isClosed hangs while listen loop is waiting on read
+
+listen :: Pipeline -> IO ()
+-- ^ Listen for responses and supply them to waiting threads in order
+listen Pipeline{..} = do
+    stream <- readMVar vStream
+    forever $ do
+        e <- try $ readMessage stream
+        var <- readChan responseQueue
+        putMVar var e
+        case e of
+            Left err -> hClose stream >> ioError err  -- close and stop looping
+            Right _ -> return ()
+
+psend :: Pipeline -> Message -> IO ()
+-- ^ Send message to destination; the destination must not response (otherwise future 'call's will get these responses instead of their own).
+-- Throw IOError and close pipeline if send fails
+psend p@Pipeline{..} message = withMVar vStream (flip writeMessage message) `onException` close p
+
+pcall :: Pipeline -> Message -> IO (IO Response)
+-- ^ Send message to destination and return /promise/ of response from one message only. The destination must reply to the message (otherwise promises will have the wrong responses in them).
+-- Throw IOError and closes pipeline if send fails, likewise for promised response.
+pcall p@Pipeline{..} message = withMVar vStream doCall `onException` close p  where
+    doCall stream = do
+        writeMessage stream message
+        var <- newEmptyMVar
+        liftIO $ writeChan responseQueue var
+        return $ readMVar var >>= either throwIO return -- return promise
 
 -- * Pipe
 
-type Pipe = Pipeline Response Message
+type Pipe = Pipeline
 -- ^ Thread-safe TCP connection with pipelined requests
 
 newPipe :: Handle -> IO Pipe
 -- ^ Create pipe over handle
-newPipe handle = Connection.fromHandle handle >>= newPipeWith
-
-newPipeWith :: Connection -> IO Pipe
--- ^ Create pipe over connection
-newPipeWith conn = newPipeline $ IOStream (writeMessage conn)
-                                          (readMessage conn)
-                                          (Connection.close conn)
+newPipe handle = newPipeline handle
 
 send :: Pipe -> [Notice] -> IO ()
 -- ^ Send notices as a contiguous batch to server with no reply. Throw IOError if connection fails.
-send pipe notices = P.send pipe (notices, Nothing)
+send pipe notices = psend pipe (notices, Nothing)
 
 call :: Pipe -> [Notice] -> Request -> IO (IO Reply)
 -- ^ Send notices and request as a contiguous batch to server and return reply promise, which will block when invoked until reply arrives. This call and resulting promise will throw IOError if connection fails.
 call pipe notices request = do
     requestId <- genRequestId
-    promise <- P.call pipe (notices, Just (request, requestId))
+    promise <- pcall pipe (notices, Just (request, requestId))
     return $ check requestId <$> promise
  where
     check requestId (responseTo, reply) = if requestId == responseTo then reply else
@@ -91,9 +175,9 @@ type Message = ([Notice], Maybe (Request, RequestId))
 -- ^ A write notice(s) with getLastError request, or just query request.
 -- Note, that requestId will be out of order because request ids will be generated for notices after the request id supplied was generated. This is ok because the mongo server does not care about order just uniqueness.
 
-writeMessage :: Connection -> Message -> IO ()
+writeMessage :: Handle -> Message -> IO ()
 -- ^ Write message to connection
-writeMessage conn (notices, mRequest) = do
+writeMessage h (notices, mRequest) = do
     noticeStrings <- forM notices $ \n -> do
           requestId <- genRequestId
           let s = runPut $ putNotice n requestId
@@ -104,8 +188,8 @@ writeMessage conn (notices, mRequest) = do
           let s = runPut $ putRequest request requestId
           return $ (lenBytes s) `L.append` s
 
-    Connection.write conn $ L.toStrict $ L.concat $ noticeStrings ++ (maybeToList requestString)
-    Connection.flush conn
+    B.hPut h $ L.toStrict $ L.concat $ noticeStrings ++ (maybeToList requestString)
+    hFlush h
  where
     lenBytes bytes = encodeSize . toEnum . fromEnum $ L.length bytes
     encodeSize = runPut . putInt32 . (+ 4)
@@ -113,13 +197,13 @@ writeMessage conn (notices, mRequest) = do
 type Response = (ResponseTo, Reply)
 -- ^ Message received from a Mongo server in response to a Request
 
-readMessage :: Connection -> IO Response
+readMessage :: Handle -> IO Response
 -- ^ read response from a connection
-readMessage conn = readResp  where
+readMessage h = readResp  where
     readResp = do
-        len <- fromEnum . decodeSize <$> Connection.readExactly conn 4
-        runGet getReply <$> Connection.readExactly conn len
-    decodeSize = subtract 4 . runGet getInt32
+        len <- fromEnum . decodeSize <$> B.hGet h 4
+        runGet getReply <$> (L.fromStrict <$> B.hGet h len)
+    decodeSize = subtract 4 . runGet getInt32 . L.fromStrict
 
 type FullCollection = Text
 -- ^ Database name and collection name with period (.) in between. Eg. \"myDb.myCollection\"